map_health_fl/web_dashboard/requirements.txt
Flask
torch
torchvision
numpy
Pillow
streamlit